Project Structure: Extract all 'handler' classes to it's own directory/namespace
