+++ Assets/PenguinQuest/Code/Controllers/Handlers/JumpUpHandler.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 using PenguinQuest.Data;
 
 
-namespace PenguinQuest.Controllers
+namespace PenguinQuest.Controllers.Handlers
 {
     [RequireComponent(typeof(Animator))]
     [RequireComponent(typeof(Rigidbody2D))]
