+++ Assets/PenguinQuest/Code/Controllers/Handlers/JumpUpHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ffdffde4918d9ad47af7f28a429ad378
+guid: 84d3f27435ee1c947b7469a39224e53b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
